prov/gni: some opts for KNL
move some functions to header files to allow for inlining.
This tends to help with KNL processor.

inline attribute in the wrong place was causing gcc
to complain.

upstream merge of ofi-cray/libfabric-cray#1096
upstream merge of ofi-cray/libfabric-cray#1100

(cherry picked from commit ofi-cray/libfabric-cray@9bfbd81633afe8ef89d0a350dc874d78f29975c9)
(cherry picked from commit ofi-cray/libfabric-cray@a76d6b24540d1d07e06e5178b9d946ea9d86fa64)
